requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-yasg==1.21.7
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.14
pytest==7.4.4
pytest-django==4.7.0
python-magic==0.4.27
gunicorn==21.2.0